improved AppNavBar
