+++ src/AwesomeApp.UI/src/frontend/views/app-nav-bar/styles.module.css
@@ -1,0 +1,3 @@
+.logoText {
+  font-weight: 700;
+}
